re-implement container creation in a simpler way
the existing atc/worker package orchestrates between the DB and
garden+baggageclaim on workers to construct containers+volumes. it
provides a `worker.Pool` for selecting a worker given a strategy, and a
`worker.Client` for running the core steps (check, get, put, task)

the current implementation has a ton of abstraction, presumably for the
sake of easier testing, but at the cost of simplicity. the tests are
also very tightly coupled to the set of abstractions, making it very
difficult to refactor - it was easier to just rewrite it in a new
package, currently atc/worker2, which no code calls at the moment. this
code is very important to be able to refactor as we work on a kubernetes
runtime

currently, I've implemented worker.FindOrCreateContainer, whose
underlying logic was pretty much copied from atc/worker, albeit without
any internal interfaces - it only depends on a garden, baggageclaim, and
the DB (as well as the pool of workers for volume streaming). I've also
implemented a new general interface for running/attaching to processes,
intended to replace the runtime.Runner (which has an interface that only
works for resources). I've also implemented a basic form of worker.Pool
that's only able to lookup volumes. tests to follow in subsequent
commits

what's changed from the original implementation:

structural
----------
* introduce more general runtime types with the intent of using them for
  other runtimes (e.g. kubernetes)
* split gardenruntime into its own package, separate from worker.Pool
* caches are no longer implemented as inputs, and cache paths are
  explicitly set in the ContainerSpec - it was easy to forget about
  them, and the original implementation deals with them in a very
  hacky/non-obvious way (if the cache isn't present on the worker,
  they're treated as remote inputs, so we create an empty volume for
  them, but they happen to not implement StreamableArtifactSource so we
  don't stream anything into them)
* code cleanup/method-call refactoring

behavioural
-----------
* resource result doesn't get saved on container properties, meaning
  re-attaching to a finished resource step won't give fetched versions
  (will fix in later commits)
* compression option is ignored, hardcoded to gzip (will fix in later
  commits)
* no p2p volume streaming (will implement in later commits)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>